requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==3.1.14
django-currencies==0.10.1
django-jazzmin==2.6.0
django-jet-reboot==1.3.3
google-api-python-client==1.4.1
httplib2==0.22.0
idna==3.4
jsonfield==2.1.1
oauth2client==4.1.3
Pillow==10.0.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
